# lab3/zad1/src/main/resources/static/order_history_service_api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/order/status/{id}": {
      "put": {
        "tags": [
          "order-post-controller"
        ],
        "operationId": "changeStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHistory"
                }
              }
            }
          }
        }
      }
    },
    "/order": {
      "get": {
        "tags": [
          "order-get-controller"
        ],
        "operationId": "getAllOrders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderHistory"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "order-post-controller"
        ],
        "operationId": "createOrderHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderHistory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHistory"
                }
              }
            }
          }
        }
      }
    },
    "/order/{id}": {
      "get": {
        "tags": [
          "order-get-controller"
        ],
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHistory"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "StatusChangeRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "OrderHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "customerName": {
            "type": "string"
          },
          "courierName": {
            "type": "string"
          },
          "deliveryStatus": {
            "type": "string"
          },
          "productNames": {
            "type": "string"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "courierName",
          "customerName",
          "deliveryStatus",
          "productNames",
          "totalPrice"
        ]
      }
    }
  }
}